cdk/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aws-hackathon-infra"
version = "0.1.0"
description = "AWS CDK infrastructure for Bedrock agents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aws-cdk-lib==2.126.0",
    "constructs>=10.0.0,<11.0.0",
    "boto3>=1.34.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "pylint>=3.0.0",
    "pyright>=1.1.0",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.pylint.messages_control]
disable = [
    "C0114",  # missing-module-docstring
    "C0115",  # missing-class-docstring
    "C0116",  # missing-function-docstring
    "R0903",  # too-few-public-methods
    "W0613",  # unused-argument
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]